+++ src/com/hermant/program/instruction/AbsoluteFloatInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.Machine;
 
 public class AbsoluteFloatInstruction extends Instruction {
 
-    AbsoluteFloatInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    AbsoluteFloatInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.ABSOLUTE_FLOAT, reg1, reg2, ramAddress);
     }
