+++ src/com/hermant/program/instruction/AddRegisterFloatInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.Machine;
 
 public class AddRegisterFloatInstruction extends Instruction {
 
-    AddRegisterFloatInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    AddRegisterFloatInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.ADD_REGISTER_FLOAT, reg1, reg2, ramAddress);
     }
