+++ src/com/hermant/program/instruction/DivideRegisterFloatInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.Machine;
 
 public class DivideRegisterFloatInstruction extends Instruction {
 
-    DivideRegisterFloatInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    DivideRegisterFloatInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.DIVIDE_REGISTER_FLOAT, reg1, reg2, ramAddress);
     }
