+++ src/com/hermant/program/instruction/EnterInstruction.java
@@ -7,7 +7,7 @@ import static com.hermant.machine.register.Register.STACK_POINTER;
 
 public class EnterInstruction extends Instruction {
 
-    EnterInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    EnterInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.ENTER, reg1, reg2, ramAddress);
     }
