+++ src/com/hermant/program/instruction/ExchangeRegisterFloatInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.Machine;
 
 public class ExchangeRegisterFloatInstruction extends Instruction implements ExchangeOperation {
 
-    ExchangeRegisterFloatInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    ExchangeRegisterFloatInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.EXAMINE_FLOAT_REGISTER, reg1, reg2, ramAddress);
     }
