+++ src/com/hermant/program/instruction/ExchangeRegisterInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.Machine;
 
 public class ExchangeRegisterInstruction extends Instruction implements ExchangeOperation {
 
-    ExchangeRegisterInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    ExchangeRegisterInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.EXCHANGE_REGISTER, reg1, reg2, ramAddress);
     }
