+++ src/com/hermant/program/instruction/LeaveInstruction.java
@@ -7,7 +7,7 @@ import static com.hermant.machine.register.Register.STACK_POINTER;
 
 public class LeaveInstruction extends Instruction {
 
-    LeaveInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    LeaveInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.LEAVE, reg1, reg2, ramAddress);
     }
