+++ src/com/hermant/program/instruction/LoadRegisterFloatInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.*;
 
 public class LoadRegisterFloatInstruction extends Instruction {
 
-    LoadRegisterFloatInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    LoadRegisterFloatInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.LOAD_REGISTER_FLOAT, reg1, reg2, ramAddress);
     }
