+++ src/com/hermant/program/instruction/LoadableInstruction.java
@@ -5,7 +5,7 @@ import com.hermant.machine.ram.RandomAccessMemory;
 
 public class LoadableInstruction extends Instruction {
 
-    public LoadableInstruction(byte code, Byte reg1, Byte reg2, Integer ramAddress) {
+    public LoadableInstruction(byte code, Byte reg1, Byte reg2, Short ramAddress) {
         super(code, reg1, reg2, ramAddress);
     }
