+++ src/com/hermant/program/instruction/MultiplyRegisterFloatInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.Machine;
 
 public class MultiplyRegisterFloatInstruction extends Instruction {
 
-    MultiplyRegisterFloatInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    MultiplyRegisterFloatInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.MULTIPLY_REGISTER_FLOAT, reg1, reg2, ramAddress);
     }
