+++ src/com/hermant/program/instruction/NegateRegisterInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.Machine;
 
 public class NegateRegisterInstruction extends Instruction {
 
-    NegateRegisterInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    NegateRegisterInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.NEGATE_REGISTER, reg1, reg2, ramAddress);
     }
