+++ src/com/hermant/program/instruction/NoOperationInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.Machine;
 
 public class NoOperationInstruction extends Instruction {
 
-    NoOperationInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    NoOperationInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.NO_OPERATION, reg1, reg2, ramAddress);
     }
