+++ src/com/hermant/program/instruction/OutputOperation.java
@@ -4,7 +4,7 @@ import com.hermant.machine.register.InstructionPointer;
 
 public abstract class OutputOperation extends Instruction{
 
-    public OutputOperation(byte code, Byte reg1, Byte reg2, Integer ramAddress) {
+    public OutputOperation(byte code, Byte reg1, Byte reg2, Short ramAddress) {
         super(code, reg1, reg2, ramAddress);
     }
