+++ src/com/hermant/program/instruction/OutputRegisterByteInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.Machine;
 
 public class OutputRegisterByteInstruction extends OutputOperation {
 
-    OutputRegisterByteInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    OutputRegisterByteInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.OUTPUT_REGISTER_BYTE, reg1, reg2, ramAddress);
     }
