+++ src/com/hermant/program/instruction/OutputRegisterCharInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.Machine;
 
 public class OutputRegisterCharInstruction extends OutputOperation {
 
-    OutputRegisterCharInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    OutputRegisterCharInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.OUTPUT_REGISTER_CHAR, reg1, reg2, ramAddress);
     }
