+++ src/com/hermant/program/instruction/OutputRegisterFloatInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.Machine;
 
 public class OutputRegisterFloatInstruction extends OutputOperation {
 
-    OutputRegisterFloatInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    OutputRegisterFloatInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.OUTPUT_REGISTER_FLOAT, reg1, reg2, ramAddress);
     }
