+++ src/com/hermant/program/instruction/OutputRegisterSignedInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.Machine;
 
 public class OutputRegisterSignedInstruction extends OutputOperation {
 
-    OutputRegisterSignedInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    OutputRegisterSignedInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.OUTPUT_REGISTER_SIGNED, reg1, reg2, ramAddress);
     }
