+++ src/com/hermant/program/instruction/PopRegisterFloatInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.Machine;
 
 public class PopRegisterFloatInstruction extends Instruction {
 
-    PopRegisterFloatInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    PopRegisterFloatInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.POP_REGISTER_FLOAT, reg1, reg2, ramAddress);
     }
