+++ src/com/hermant/program/instruction/PushFlagsInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.Machine;
 
 public class PushFlagsInstruction extends Instruction {
 
-    PushFlagsInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    PushFlagsInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.PUSH_FLAGS, reg1, reg2, ramAddress);
     }
