+++ src/com/hermant/program/instruction/PushRegisterFloatInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.Machine;
 
 public class PushRegisterFloatInstruction extends Instruction {
 
-    PushRegisterFloatInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    PushRegisterFloatInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.PUSH_REGISTER_FLOAT, reg1, reg2, ramAddress);
     }
