+++ src/com/hermant/program/instruction/PushRegisterInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.Machine;
 
 public class PushRegisterInstruction extends Instruction {
 
-    PushRegisterInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    PushRegisterInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.PUSH_REGISTER, reg1, reg2, ramAddress);
     }
