+++ src/com/hermant/program/instruction/SubtractRegisterFloatInstruction.java
@@ -4,7 +4,7 @@ import com.hermant.machine.Machine;
 
 public class SubtractRegisterFloatInstruction extends Instruction {
 
-    SubtractRegisterFloatInstruction(Byte reg1, Byte reg2, Integer ramAddress) {
+    SubtractRegisterFloatInstruction(Byte reg1, Byte reg2, Short ramAddress) {
         super(Instruction.SUBTRACT_REGISTER_FLOAT, reg1, reg2, ramAddress);
     }
